feat: implement comprehensive architecture refactor
This commit introduces a complete architectural restructure to establish
a proper dual-layer frontend system with clear separation of concerns:

## Server Layer Reorganization
- Move core server logic to src/server/ (env.ts, index.ts)
- Reorganize API adapters to src/api/ (format.ts, stream.ts, providers.ts)
- Relocate templates to src/templates/ (index.ts, privacy.ts, terms.ts)
- Add image proxy route and tests

## Client Architecture Implementation
- Establish modular client architecture in src/client/
- Implement domain-specific modules (bestPractices, howToImplement, etc.)
- Create shared components with proper inheritance patterns
- Add comprehensive type definitions and service layers

## Build System Integration
- Update generated bundles for new module structure
- Maintain compatibility with existing Cloudflare Workers deployment
- Ensure proper asset organization and template compilation

## Benefits
- Clear separation between development (src/client) and runtime (modules)
- Improved maintainability through modular architecture
- Enhanced type safety across all client modules
- Better testability with structured service layers
- Scalable foundation for future feature development

This refactor maintains full backward compatibility while establishing
a robust foundation for continued development.
